refactor: move presentation entries and build
