+++ extension/src/app/presentation/content/index.ts
@@ -1,6 +1,6 @@
-import { createSignIn, type PendingSignIn } from "../app/application/auth/sign-in";
-import { createChromeTokenStore } from "../app/infrastructure/providers/chrome-token-store";
-import { pollForToken, requestDeviceCode } from "../app/infrastructure/providers/github-device-flow";
+import { createSignIn, type PendingSignIn } from "../../application/auth/sign-in";
+import { createChromeTokenStore } from "../../infrastructure/providers/chrome-token-store";
+import { pollForToken, requestDeviceCode } from "../../infrastructure/providers/github-device-flow";
 import {
   render,
   renderError,
@@ -17,15 +17,15 @@ import {
   type SemanticDiffResponse,
   targetKey,
   unresolvedRemaining,
-} from "../app/domain/diff/types";
-import { must } from "../app/presentation/util/must";
-import { createAuthRetries } from "../app/application/overlay/auth-retries";
-import { createFileView } from "../app/application/overlay/file-view";
-import { createViewRegistry, type ViewEntry } from "../app/application/overlay/views";
-import { createViewState, type ViewState } from "../app/application/overlay/view-state";
-import type { View } from "../app/application/overlay/view-mode";
+} from "../../domain/diff/types";
+import { must } from "../util/must";
+import { createAuthRetries } from "../../application/overlay/auth-retries";
+import { createFileView } from "../../application/overlay/file-view";
+import { createViewRegistry, type ViewEntry } from "../../application/overlay/views";
+import { createViewState, type ViewState } from "../../application/overlay/view-state";
+import type { View } from "../../application/overlay/view-mode";
 import { type DiffPage, type FileEntry, parseDiffUrl, parsePrPage, scanUnityFiles } from "./detect";
-import { fillDeviceCode } from "./devicePage";
+import { fillDeviceCode } from "./device-page";
 import { createToggle, type Toggle } from "./toggle";
 
 const ERROR_TEXT: Record<BackgroundError, string> = {
