+++ extension/src/app/presentation/content/toggle.test.ts
