+++ extension/src/app/presentation/renderer/icons.ts
