+++ extension/src/app/presentation/renderer/styles.ts
